requirements.txt
Flask
flask-cors
python-dotenv
Pillow
google-generativeai
opencv-python
numpy
gunicorn